+++ docs/CHANGELOG.md
@@ -3,6 +3,13 @@ All notable changes to this project will be documented in this file.
 
 For more details on new features, please check the [Manual](./MANUAL.md).
 
+<details><summary>7.2.0 - 8 February 2023</summary>
+
+### Added
+- New `[img2pez]` shortcode (Hard Prompts Made Easy)
+
+</details>
+
 <details><summary>7.1.0 - 7 February 2023</summary>
 
 ### Added